refactor: extract timegrid event resizing logic to hook
